pyproject.toml
[project]
name = "mcp-docx-server"
version = "0.1.0"
description = "MCP server for manipulating Word documents"
dependencies = [
    "mcp[cli]",
    "python-docx",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
